requirements.txt
alembic==1.5.8
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.63.0
FastAPI-SQLAlchemy==0.2.1
greenlet==1.0.0
h11==0.12.0
idna==3.1
Mako==1.1.4
MarkupSafe==1.1.1
passlib==1.7.4
psycopg2==2.8.6
pydantic==1.8.1
PyJWT==2.0.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-multipart==0.0.5
six==1.15.0
SQLAlchemy==1.4.3
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.13.4
